+++ ios/flutter_in_store_app_version_checker/Sources/flutter_in_store_app_version_checker/SwiftInStoreAppVersionCheckerPlugin.swift
@@ -2,9 +2,11 @@ import Flutter
 import UIKit
 
 public final class SwiftInStoreAppVersionCheckerPlugin: NSObject, FlutterPlugin {
+  private static let channelName = "github.com/ziqq/instoreappversionchecker/app_metadata"
+
   public static func register(with registrar: FlutterPluginRegistrar) {
     let channel = FlutterMethodChannel(
-      name: "github.com/ziqq/flutter_in_store_app_version_checker",
+      name: channelName,
       binaryMessenger: registrar.messenger()
     )
     let instance = SwiftInStoreAppVersionCheckerPlugin()
@@ -12,6 +14,18 @@ public final class SwiftInStoreAppVersionCheckerPlugin: NSObject, FlutterPlugin 
   }
 
   public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
-    result("iOS " + UIDevice.current.systemVersion)
+    switch call.method {
+    case "getAppMetadata":
+      result([
+        "packageName": Bundle.main.bundleIdentifier,
+        "version": Bundle.main.object(
+          forInfoDictionaryKey: "CFBundleShortVersionString"
+        ) as? String,
+      ])
+    case "getPlatformVersion":
+      result("iOS " + UIDevice.current.systemVersion)
+    default:
+      result(FlutterMethodNotImplemented)
+    }
   }
 }