+++ test/util/mocks.dart
@@ -2,14 +2,10 @@
 
 import 'package:http/http.dart' as http;
 import 'package:mockito/annotations.dart';
-import 'package:package_info_plus/package_info_plus.dart';
 
 export 'mocks.mocks.dart';
 
-@GenerateMocks([], customMocks: [
-  MockSpec<http.Client>(),
-  MockSpec<PackageInfo>(),
-])
+@GenerateMocks([], customMocks: [MockSpec<http.Client>()])
 void getGenerateMocks() {}
 
 const fakeAndroidBuildVersion = <String, dynamic>{